Be able to delete user
```
$ vim app/controllers/users_controller.rb
$ vim app/views/users/index.html.haml
$ vim app/views/users/show.html.haml
```

+++ app/controllers/users_controller.rb
@@ -24,6 +24,10 @@ class UsersController < ApplicationController
 
   def destroy
     user = User.find(params[:id])
+    if user && user.admin
+      redirect_to users_url, notice: 'fail'
+      return
+    end
     user.destroy
     redirect_to users_url, notice: 'User was successfully destroyed'
   end
